+++ neoforge/src/main/java/net/mehvahdjukaar/moonlight/neoforge/MoonlightForgeClient.java
@@ -2,19 +2,13 @@ package net.mehvahdjukaar.moonlight.neoforge;
 
 
 import net.mehvahdjukaar.moonlight.api.entity.IControllableVehicle;
-import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
 import net.mehvahdjukaar.moonlight.api.platform.configs.ModConfigHolder;
 import net.mehvahdjukaar.moonlight.core.MoonlightClient;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.Input;
-import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.repository.FolderRepositorySource;
-import net.minecraft.server.packs.repository.PackSource;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.level.levelgen.structure.PoolElementStructurePiece;
 import net.neoforged.bus.api.EventPriority;
 import net.neoforged.bus.api.IEventBus;
-import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.event.lifecycle.FMLLoadCompleteEvent;
 import net.neoforged.neoforge.client.event.MovementInputUpdateEvent;
@@ -22,10 +16,7 @@ import net.neoforged.neoforge.client.event.TextureAtlasStitchedEvent;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import net.neoforged.neoforge.common.NeoForge;
-import net.neoforged.neoforge.event.AddPackFindersEvent;
 import net.neoforged.neoforge.event.entity.player.ItemTooltipEvent;
-
-import java.nio.file.Path;
 
 public class MoonlightForgeClient {
 
@@ -33,7 +24,7 @@ public class MoonlightForgeClient {
         modEventBus.addListener(MoonlightForgeClient::afterLoad);
         modEventBus.addListener(EventPriority.LOWEST, MoonlightForgeClient::onTextureStitch);
         NeoForge.EVENT_BUS.addListener(MoonlightForgeClient::onInputUpdate);
-        NeoForge.EVENT_BUS.addListener( EventPriority.LOWEST, MoonlightForgeClient::itemTooltipEvent);
+        NeoForge.EVENT_BUS.addListener(EventPriority.LOWEST, MoonlightForgeClient::itemTooltipEvent);
     }
 
     public static void itemTooltipEvent(ItemTooltipEvent event) {